requirements.txt
openai
langchain
python-dotenv
langchain-openai
pypdf
langchain-community
beautifulsoup4
faiss-cpu
langchain-text-splitters
langchainhub
pandas
yfinance==0.2.54
matplotlib
python-dotenv
pytest
